test: Drop check_axe()
We have dragged this around for many years without anyone having the
time or interest to finish it, or us ever looking at the results. The
new version also makes the browser hang.

So let's remove this for now. We can easily restore it if/when we need
it.

Fixes #20022