kdump: use a temporary file for QUnit test
With the fsinfo channel this test fails because it uses an relative path
which is unsupported in fsinfo. As using a relative path with a
hardcoded filename means we always write to the same file if you
repeatedly run the test which could lead to undesired
side-effects/issues

Use QUnit's before/after helper functions to always create a temporary
file, in the future `base1/test-file.js` should also use this instead of
the current hack.