accounts: Resolve datarace issue when creating accountsInfo
On account creation /etc/passwd and /etc/shadow race so if passwd is
parsed first, then we would end up with the wrong state as lastlog was
not parsed yet. Fix this issue by looking up the accounts from details,
instead of relying on the ordering of those lists.

This also adds support for lastlog2. Lastlog2's output differes from
lastlog so cockpit cannot no longer rely on the odering of /etc/passwd