logs: Do not use duplicate object to keep state
The source of the truth should be `currentIdentifiers` as defined in `logs.jsx`.
`logsJournal.jsx` had separate object to keep the same information and
than this `currentIdentifiers` was being synced. This was not ideal
solution which might fail when there are multiple updates at the same
time as well as it might not trigger update correctly when just editing
set object.

This commit only uses one source of the truth and uses updater callback
to make sure the list is always correctly updated.