lib: introduce an upload() helper for fsreplace1
Introduces a helper function which handles uploading a ReadableStream to
a server with `fsreplace1`, sending data in chunks, with proper flow
control and progress reporting.  This is made to be used for
cockpit-files but is re-usable for other plugins.

Add a React component to the playground as a proof of concept, and add a
couple of integration tests based on it.

Co-authored-by: Allison Karlitskaya <allison.karlitskaya@redhat.com>