lib: Avoid infinite layout measurements in plots
The SvgPlot component forces a re-render when the size of its outer
"div" has changed. This requires that this outer div settles on a
stable size.

However, some browsers can optionally add noise to the result of
getBoundingClientRect, which causes a constant re-rendering of the
plots since the "div" never seems to settle.

Thus, we tolerate the noise and only re-render when the size has
changed significantly.

Fixes #20171