tools: add 'release' script
This is a new script for building Cockpit releases.  It essentially does
what we do now, except:

 - we use the container version specified in our .cockpit-ci/container
   file

 - we download everything with a single invocation of `git clone`, which
   downloads all the required submodules, but only does a shallow fetch

 - we run the build process entirely offline, inside of its own
   container

The new script is meant to be more accessible to anyone who may be
interested in running it, including outside of GitHub, and is written in
a way that attempts to convince its reader that nothing tricky is going
on.  It should always produce a byte-for-byte identical release as the
official one.  It acts as a convenient way to verify the reproducibility
and validity of a Cockpit release, and also as an alternative to
downloading the tarball from GitHub.