bridge: tweak a comment in AsyncChannel
6ddf67698ee733e8bb6fcc93df8a6c9f68e52ce0 changed the EOF return value of
`AsyncChannel.read()` from `b''` to `None` to account for the
possibility that the user might have sent an empty message (which has a
semantic meaning in at least one other channel — `fsreplace1`).
Unfortunately, it also introduced a comment documenting the old return
value behaviour.

Update the comment to reflect the new reality.