bridge: rewrite fsreplace1 as AsyncChannel
Our previous reason for avoiding tempfile.NamedTemporaryFile() was that
we didn't want to chown() after the fact, but we've crossed that bridge
already since b9a0cdcc41bf02a1864bb3f51e9eafbb46c61d9f, so we may as
well go all the way.

Rewrite things as a single run() function using NamedTemporaryFile as a
context manager.  This is a fairly substantial simplification.  We also
make the channel properly asynchronous by dispatching the actual write
to an executor.  We also add fdatasync(), which we were missing before.

We need to make a minor adjustment to AsyncChannel to support sending
close arguments, similar to the way we handle them on GeneratorChannel.