test: workaround for watch channel issue
The watch channel can emit events even after `.close()` is called, which
can cause issues in the `kdump/test-config-client.js` QUnit test if the
timing of the file write changes (which it is about to do, in a upcoming
commit which reimplements fsreplace1 asynchronously).

Removing the event listener is a sure way to ensure we won't get further
events.