test: Use `nft` instead of `iptables`
`iptables` has been a backwards compatibility shim for nftables for a
long time now, and finally does not work on CentOS/RHEL 10 any more.

The first case in TestConnection.testBasic takes the socket away on the
login page with a TCP reset reply. We can get the same effect by just
stopping the socket.

The second case silently drops the packets to cover the "reply timeout"
functionality of the shell. Replace `iptables` with the corresponding
`nft` commands. It would be nice to use firewalld for that, but that has
a default rule to always allow packets on established connections -- but
severing that is exactly the goal of this test.