test: use f-strings in asserts in test-format
Use our fancy new tagged templates to name our assertions in
test-format.  In addition to being a lot easier (and shorter) to write,
it means we see `{"precision": 2}` instead of `[object Object]`.

Also, leave out the value that we're comparing it equal to.  This value
is already displayed on assertion failures, and having it shown twice
only makes the output harder to read.