bridge: use tempfile.mkstemp() for fsreplace1
If we could depend on Python 3.12 and use delete_on_close=False then
using NamedTemporaryFile as a context manager would be really useful.
Since we can't, and already have to deal with the deletion manually,
let's also just deal with the close part manually.  Getting an fd and a
filename fits our uses a lot better anyway: most things we call want to
operate on an fd and having the name in a variable lets us easily flag
the cases where we don't want to try to delete it.

We move the `finally:` handling of closing the fd and (conditionally)
removing the temporary file to a separate internal `try:` block.  This
has nothing to do with the delete case (which does want to share the
error handling, which is now the only thing the outer `try:` does).