lib: Generalize manifest config reading
We already use that on the storaged and apps pages, and going to add it
to metrics as well. As storage and apps want to look at different
os-release fields, start with a very generic "matchlist" API. We may later add
an async convenience wrapper which reads os-release by itself and matches on
common fields.

While at it, factorize the duplicated config reading in apps to a
get_packages() helper, consistently use `async`, and split the imports
into "library" and "page specific".