fmf: Reorganize and rebalance tests
The "basic" vs. "optional" package split was only a RHEL 8 thing, and
was removed in commit 7a372ddd. Also the test plans are very imbalanced:
basic takes 14 mins, network only 4.5 mins, and optional about 30 mins.

Reorganize them to split the storage tests in two plans
(`storage-{basic,extra}`) and merge the others into a `main`.

As we now have to enumerate test class names explicitly, this cleans up
the `EXCLUDES` list quite a bit: We can drop the parts where we skipped
all tests of a class, and instead just not mention the class in `TESTS`.