test: Port non-text Browser.key_press() to Browser.key()
That makes the code easier to read/understand, and paves the way for
deprecating key_press(). Add a `modifiers` kwarg to it too, as we need
that for at least TestTerminal.

Unfortunately Chromium is a bit buggy and doesn't understand some `key=`
names like "Backspace" or "ArrowDown" in some (but not all!)
circumstances/tests. It works fine for most other, like "Tab", or
(really curiously) "ArrowLeft". Add a workaround by translating these to
`windowsVirtualKeyCode`, as `key_press()` does.