test/static-code: make mypy deps configurable
We hardcode a check for systemd_ctypes, ferny, and beiboot in
test/static-code and skip running mypy if they're not checked out, but
that doesn't make sense for the other projects which use this script.

Move the configuration to a separate file, so each package can provide
their own.  If no file is present, there are no dependencies: this is
expected to be the normal case for starter-kit-based projects.