pkg/lib: make the upload-helper WebKit-compatible
Our use of `{ mode: 'byob' }` gets us in trouble on WebKit.  We see
"ReadableStreamBYOBReader needs a ReadableByteStreamController" on any
attempt to upload a file.  This affects:

 - Cockpit Client
 - Epiphany
 - Safari on macOS and iOS

And indeed, https://caniuse.com/mdn-api_readablestreambyobreader agrees
that we can't use this yet.

We also can't simply change to the default reader: the reason we chose
the BYOB reader is that it let us control the block size that was
returned.  The default reader returns large blocks (on the order of 2MB)
which we can't send over the websocket, because of limitations in
`cockpit-ws`:

  server is trying to frame of size 524294 or greater, but max supported size is 128KiB
  Received extremely large WebSocket data from the server

It seems we can't get around the fact that we're going to have to do
some slicing here.  Let's try to use the BYOB API if possible, and fall
back to an adaptor if we can't.

Fixes #20615