test: Drop Grafana selector class names
The latest Grafana version in the services refresh [1] replaced all
human readable and sensible CSS classes with unpredictable junk like
"css-140ea9t". So stop relying on the classes and just ensure that the
"Load average" panel contains "minute" and any floating point number.
This works with both the current and the new image.

[1] https://github.com/cockpit-project/bots/pull/6483