bridge: Split Channel.send_data() into send_{bytes,text}()
This factorizes a few `.encode()` calls, and clarifies which callers are
guaranteed to send UTF-8 text, and which can send arbitrary bytes.

This leaves AsyncChannnel and GeneratorChannel as the places which
potentially send non-UTF8-data to text channels. We'll deal with them in
the next commit to not change too many things at once.

Move `send_json()` below `send_text()` to keep all `send_*()` methods
together.