pkg/lib: add `Dialogs.run()`
Replace the existing `await Dialogs.show()` approach with a new function
called `Dialogs.run()`.  This instantiates the passed component type
with the given attributes, plus one additional attribute: `dialogResult`,
which has `.resolve()` and `.reject()` methods on it which determine how
the promise returned by `Dialogs.run()` will be resolved.

In addition to simplifying the implementation, this keeps the data flow
local and lets us ensure that the return value of `.run()` has the same
type as the value passed to `.resolve()`.

`Dialogs.show()` remains, but is no longer awaitable: there is only a
single user of the Promise, in Cockpit Files, and it already has an open
PR — https://github.com/cockpit-project/cockpit-files/pull/585 — to port
to the new API.