tsconfig: switch target to es2020
This is the base level of the language that the compiler targets (or
understands).  We're about to add some code which uses `#private`
instance variables, which is a es2017 feature.  We already depend on
es2020 library features, so let's allow the same for the core language.

This has a side-effect: setting `target: 'es2020'` changes the default
module resolution approach.  Force it back to `"node"` so we can still
find our `node_modules`.