cockpit.js: tweak "factory" setup.
Move the creation of the `cockpit` object from the toplevel and into the
factory function.  Assign the toplevel as the result of the function,
instead of having the factory modify the toplevel.

In addition to being conceptually cleaner, it also makes TypeScript
happier: we no longer get a complaint on every assignment, and in fact,
it now recognises the pattern (create an objetct and assign methods to
it) and infers a fairly decent implicit type for the return value of the
function (and therefore) the `cockpit` object.