fsinfo: use `new Channel` directly
Instead of calling into `cockpit.channel()` which is just a wrapper
around `new Channel()`, use it directly.  Also use our JSON helpers from
`common.ts` instead of `cockpit.d.ts`.

This removes our dependency on `cockpit.js` in the `fsinfo` code, which
is helpful, since we're about to introduce a dependency in the other
direction.

The typing is a bit sideways here: TypeScript has no way to know that
the return value of cockpit.channel() (which we've annotated) is trying
to describe the Channel class (which is still in javascript, and
unannotated).