pkg/lib/_internal/common.ts: add some utility types
Use this as a convenient location for our JSON helper types, and also
for our augmentations to the `window` global — it doesn't matter where
those live, as long as they're included in the compilation, and this
file gets (transitively) included by just about everything.

We can drop index.d.ts now.