systemd: Change "Uptime" duration to "Up since" point in time
In a following commit we will change timeformat.distanceToNow() to drop
support for formatting a duration. The uptime is the only problematic
consumer in all our projects which grammatically requires a duration.

Change it to a point in time ("... hours ago") and thus relabel it from
"Uptime" to "Up since".