storaged: Change "Runtime" duration to "Started" point in time
Similarly to the previous commit, change the process start time in the
"things that keep a block device busy" dialog section to a point in
time.

The only remaining format_delay() consumer is for the Jobs panel, which
is a future time -- so temporarily add an `addSuffix` option to match
`timeformat.distanceToNow()`.