test: robustify creating a partition table in a test
In our tests the format dialog opens while /mnt is just unmounted and
storaged not having had the time yet to process the UDisk event that
the filesystem is not mounted.

As our UI does not show the mounted status in the overview, retry
opening the dialog and try to format an empty partition table.
To ensure we test that this actually succeeds we know assert that the
disk changed from ext4 filesystem to an "empty drive".