test: Fix potential race condition in TestStorageswap
Like everything else, stopping/starting a swap partition is
asynchronous, and doesn't immediately reflect in /etc/fstab. This check
only worked because waiting for the first swap usage metric takes a
while and requires the swap partition to be in the desired state.

But without PCP (which we are going to disable on debian-testing) this
isn't reliable, so add a waiting loop to make the fstab check
independent from the metrics check.