pkg/lib: move methods out of Transport.constructor
These methods no longer depend on anything inside of the constructor()
function and can be moved out.  This is safe because there are no call
sites that use these functions as callbacks, so `this` will always be
correctly bound.

The methods formerly exported on the `self` object are turned into
public methods and the others are turned into privates with `#`.

The one tricky case here is `dispatch_data()` which is now a public
method, but also used as the WebSocket `.onmessage` handler.  In that
case we use an arrow to make sure `this` is properly provided:

    this.#ws.onmessage = event => this.dispatch_data(event);