test: Complete and clean up check-shell-multi-os
This only tested connecting from dev to stock, but not the other way around.
That's equally important for ensuring forwards compatibility in the bridge and
ws protocols. Add that direction as well, both direct login and via "Add host"
in the Shell.

TestRHEL8 and TestMultiOSDirect were using the same expensive VM setup. Run all
scenarios in a single test, i.e. eliminate the latter class.

This test used a lot of redundant custom code for adding a machine via the
Shell. Use our testlib helpers instead.

Move the test to run against RHEL 8.10. That is the latest, and
long-term supported RHEL 8 version.