tests: Reduce dependency on #host-toggle
`#hosts-sel` will stay for showing the current user/host, but the toggle
will go away by default when we deprecate the host switcher.

The only remaining reference to it is in `testKeyboardNavigation` now,
which is intrinsically necessary. That test will have to be adjusted
when we actually disable the host switcher.