pkg/lib: eliminate "out" filtering
"in" filtering is used by the shell to inspect all incoming messages
for forwarding to the appropriate frame, but nothing uses "out" filters,
outside of a single test case which exists only to test the feature.

Removing this will substantially simplify the code and allow for further
cleanups, so let's drop it.