pytest: skip pidfd ENOSYS on Python 3.12+
Since Python 3.12, we've been getting a deprecation warning about our
use of SafeChildWatcher, which is due for removal in Python 3.14.  We
only do that if we are running on a kernel without pidfd support (which
is unfortunately still the case on RHEL systems).

We effectively test this code path on our tox runs for Python versions
3.9 (where PidfdChildWatcher was added) through 3.11 (the last version
before SafeChildWatcher was deprecated), so stop testing on 3.12+, to
avoid the warning (and the error, once support is finally dropped).