pytest: fix pytest-asyncio version check
This is a fixup for f3215c0163bde8e6f6e014d51c7e10fa40f0d747.

We used a check for Python 3.6 to see if we had an old version of
pytest-asyncio.  This was a bad idea, because Debian has old
pytest-asyncio even though it has a newer Python.

Change the check to directly check the version of pytest-asyncio.

Version 0.22 is the first version to have the event_loop_policy fixture
and also the version that the deprecation warning was introduced.