test: Clean up Browser.select_PF[45]()
We don't have PatternFly 4 anywhere any more, so the naming of these was
confusing. select_PF4() is also not used any more in external cockpit-*
projects.

Replace these with a new select_PF() (with fixed typing) that applies to
a current PatternFly `<Select>`. Also provide a select_PF_deprecated()
variant for the deprecated one (which we still use exclusively in
cockpit, but e.g. not any more in c-machines).