lib: Provide reduced SSH host dialogs API as library
This makes the dialog re-usable for frames who want to do their own
remote host handling, and allows us to eventually disable the host
switcher in the shell.

The existing code in the shell makes deep assumptions about the
`machines_ins` (aka. pkg/shell/machines/machines.js) API, none of which
we want to support. Also, it's rather deeply convoluted. So reimplement
a simple "cockpit-connect-ssh" library API with re-using the design and
React parts of pkg/shell/hosts_dialog.jsx. But remove the "auto-unlock
SSH key" functionality -- it's not good practice to advertise that, it
depends on functionality of cockpit-ws which we want to get rid of
(retain password for some time after login) and it doesn't make much
sense or even work for this API.

We will develop the shell and lib parts independently from now on, and
hopefully eventually drop the shell feature.

However, we can share the SSH key helper scripts, so move these to
pkg/lib as well.

Add a "remote channel" playground page for testing this. Test all the
functionality in the new /check-ssh-api. Add TestMultiMachine tests for
testing that functionality from a direct SSH login, and a secondary ssh
shell session.

https://issues.redhat.com/browse/COCKPIT-1144