css: Shuffle some code around to make SASS happy
We're supposed to have CSS related to the top level at the top anyway, but SASS is getting picky about it.

I've fixed the issues in the code, and this makes the SASS compiler happy with our SCSS.

This *shouldn't* change anything.

The warning from SASS was:

> ▲ [WARNING] Sass's behavior for declarations that appear after nested
rules will be changing to match the behavior specified by CSS in an upcoming
version. To keep the existing behavior, move the declaration above the nested
rule. To opt into the new behavior, wrap the declaration in `& {}`.