build: rename fail.html.c to fail-html.c
We hardcode the failure page into cockpit-ws because we want to be able
to serve it even in situations where the install is totally broken.  We
do that by use of a very trivial script which escapes the html file as a
string in a temporary .c file which gets included as part of cockpit-ws.

Unfortunately, the name of that file — fail.html.c — was poorly chosen.

make has a built-in implicit rule for turning %.c into % by running the
C compiler.  When it sees that fail.html is older than fail.html.c it
concludes that the correct thing to do is to bring it up to date by
compiling it, replacing the original HTML page with a binary.  That's
wrong of course, and doesn't work anyway, since there's no main().

The only reason that this hasn't been a problem up to now is that there
seems to be some complicated side-effects caused by having .html
included in our TEST_EXTENSIONS list, but the next commit wants to
remove that.

Let's just avoid the problem up by calling the file by a different name
that won't match that implicit rule.

While we're at it, remove a single stray tab character that somehow made
its way into the fail page.