storage: Drop obsolete `pointerEvents: none` hack in controls
This was introduced in commit 2531e0eb3e mostly for cockpit-machines.
Since then most of these hacks disappeared, but there's two left.

`StorageControl` avoids that problem by wrapping the `Tooltip` contents
in an extra span, so that tooltips work for disabled elements.

This avoids a race condition: When the `excuse` is empty, but there's a
spinner, then the button is not formally disabled, but clicks go into
the void. This fixes the flaky TestStorageLuksDestructive.testLuks1Slots.