cockpit.js: fix cockpit.file().replace('')
We call iterate_data() on whatever the user passed to us, but this
function will make zero callbacks if the string is empty.  That's a
problem, because the fsreplace1 channel interprets this as a request to
delete the file.

If we have an empty string, then send the empty string to make sure we
end up with an empty file.  Also treat the `null` case more explicitly
(and add a type annotation for it: we missed that before).

Add a test case too.