test: Re-teach Browser.input_text() about \n
`Browser.input_text()` is not generally meant for non-ASCII characters
(that's `Browser.key()`). But a newline is such a common case that it's
worth handling it speciallly.

This was originally introduced in commit c3f903ed547b05b but I missed
that in the BiDi conversion in commit 6f23d2750d1fdf.