beiboot: support for temporary UserKnownHostsFile
In the bastion host case, create a temporary UserKnownHostsFile for ssh
and populate it with data from the authorization (if provided).  On
successful login, send the updated key data back to the user via the
long-disused `x-login-data` mechanism.  That'll get it added to the JSON
blob that is sent as the reply to the `GET /login` request. On failed
login, send the key data as part of the Cockpit protocol error, in a
`known-hosts` attribute.

Co-Authored-By: Martin Pitt <mpitt@redhat.com>