Revert "test: Increase RAM for TestKdumpNFS machine"
We have a better theory now of how kdump works and especially what
crashkernel means, and increasing the physical RAM of a machine is not
expected to help with growing initramfs sizes.

It _did_ help, for mysterious reasons, but not totally reliably. So
let's revert this, save some RAM, and wait for a real fix.

See https://bugzilla.redhat.com/show_bug.cgi?id=2305968

This reverts commit f8e7e23c0175ce9a6a149f9229d3334f1c8a18d3.