src/build_backend: sort file list for archives
We use os.walk() to list files on the filesystem which may be
interesting for including in our sdist, wheel, and beipack.  This
returns the list of a files in a non-deterministic order.

Before we add the files to the sdist or the wheel (which the beipack is
built from), make sure to sort the list.